# contracts/openapi/orders.yaml
openapi: 3.1.0
info:
  title: Waypoint Order Service
  version: 0.2.0
servers:
  - url: /api/v1
paths:
  /orders:
    get:
      operationId: listOrders
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: status, in: query, schema: { type: string } }
        - { name: brand, in: query, schema: { type: string } }
        - { name: outlet_id, in: query, schema: { type: string } }
        - { name: requested_delivery_date, in: query, schema: { type: string, format: date } }
      responses:
        "200":
          description: Order list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Order" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      operationId: createOrder
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateOrder" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  order: { $ref: "#/components/schemas/Order" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/Internal" }
...
  /orders/{id}/tracking:
    get:
      operationId: getOrderTracking
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/ID" }]
      responses:
        "200": { description: Aggregated order, planning, delivery, and receipt state }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema: { type: string }
  schemas:
    CreateOrder:
      type: object
      required: [requestedDeliveryDate, orderUnits, orderWeightKg, orderVolumeM3, temperatureRequirement]
      properties:
        requestedDeliveryDate: { type: string, format: date }
        orderUnits: { type: integer, exclusiveMinimum: 0 }
        orderWeightKg: { type: number, exclusiveMinimum: 0 }
        orderVolumeM3: { type: number, exclusiveMinimum: 0 }
        temperatureRequirement: { type: string, enum: [ambient, chilled] }
    Order:
      type: object
      properties:
        id: { type: string }
        orderRef: { type: string }
        outletId: { type: string }
        brand: { type: string }
        requestedDeliveryDate: { type: string }
        orderUnits: { type: integer }
        orderWeightKg: { type: number }
        orderVolumeM3: { type: number }
        temperatureRequirement: { type: string }
        status: { type: string }
        createdBy: { type: string }
...
  responses:
    BadRequest:
      description: Invalid request
    Unauthorized:
      description: Unauthenticated
    Forbidden:
      description: Not authorized
    NotFound:
      description: Missing
    Internal:
      description: Unexpected error